Add files via upload
The project is from US based financial institutions, where the main idea is to find whether the consumer has disputed or not after raising a complaint. One of the reasons the consumer might have disputed is timely response from the company. The normal TAT (Turn Around Time) to solve a complaint is 3 days. If company is taking more than the TAT, then consumer might dispute.

The data set consists of  555957 rows and 18 columns out of which 17 columns were of Object data type and only 1 column was of int32 data type.